+++ .env
@@ -8,6 +8,13 @@ AWS_SECRET_ACCESS_KEY = s8H4p8ZDVEPVLn5/VFJ3gBMP+EYP1BQqrnPBkHK5
 
 AWS_S3_BUCKET_NAME = resolvia
 
+AWS_DEFAULT_REGION = ap-south-1
+
 CLOUDFRONT_BASE_URL = https://d37m1ffxg00bok.cloudfront.net
 
+S3_COMPLETE_BASE_URL = https://resolvia.s3.ap-south-1.amazonaws.com
+
 S3_BASE_URL = https://resolvia.s3.amazonaws.com
+
+
+
