+++ app/routes/resolve.routes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const verifyToken = require('../middleware/verifyUser.js');
-const { create, uploadFile,  getResolves} = require('../controllers/resolve.controller.js');
+const { create, getResolves } = require('../controllers/resolve.controller.js');
 const multer = require('multer');
 
 const storage = multer.memoryStorage();
@@ -11,11 +11,9 @@ const upload = multer({
   },
 });
 
-
 const router = express.Router();
 
-router.post('/create', verifyToken, create);
-router.post('/upload', upload.single('document'), uploadFile);
+router.post('/create', verifyToken, upload.single('document'), create);
 router.get('/getresolves', getResolves);
 
 module.exports = router;
